Rename parsed_fields to replace
