+++ src/main.rs
@@ -8,7 +8,7 @@ mod app;
 mod app_runner;
 mod fields;
 mod logging;
-mod parsed_fields;
+mod replace;
 mod tui;
 mod ui;
 mod utils;
